Use the correct python interpreter to run repoman
Repoman has a shebang that hardcodes /usr/bin/python, so if we run
the repoman executable directly, the python version set in .travis.yml
is ignored. To fix this we now run repoman through "python", which
uses PATH lookup and therefore finds the correct python interpreter
in the travis virtualenv installation.