Fix symlink bug and update Portage version
A symlink is made from the official Portage repos.conf
to repos.conf in /etc/portage, however an absolute path is not used.
This means that the symlink has a relative path and does not point
to the intended location.  This commit fixes that by using an absolute
path for the symlink, producing the intended result.

Also, update the PORTAGE_VER to 2.3.17